End of Section 18

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 
@@ -19,7 +20,7 @@ pageSize = 5;
 loading=false;
 
 
-  constructor(private messageService:MessageService) { }
+  constructor(private messageService:MessageService, private confirmService:ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessage();
@@ -36,8 +37,12 @@ loading=false;
   }
 
   deleteMessage(id:number){
-    this.messageService.deleteMessage(id).subscribe(()=>{
-      this.messages.splice(this.messages.findIndex(m=>m.id==id), 1)
+    this.confirmService.confirm('Confirm Delete Mesage', 'this connot be undone').subscribe(result=>{
+      if(result){
+        this.messageService.deleteMessage(id).subscribe(()=>{
+          this.messages.splice(this.messages.findIndex(m=>m.id==id), 1)
+        })
+      }
     })
   }
